+++ ScreenshotsParseTool/scripts/start_dark.py
@@ -1,12 +1,15 @@
+import os
 from PyQt6 import QtCore, QtGui, QtWidgets
 
 
 class Ui_Dialog(object):
     def setupUi(self, Dialog):
+        self.current_dir = os.path.dirname(__file__)
+        self.path = os.path.abspath(os.path.join(self.current_dir, os.pardir))
         Dialog.setObjectName("Dialog")
-        Dialog.resize(320, 240)
+        Dialog.setFixedSize(320, 240)
         icon = QtGui.QIcon()
-        icon.addPixmap(QtGui.QPixmap("app_images/logo.png"), QtGui.QIcon.Mode.Normal, QtGui.QIcon.State.Off)
+        icon.addPixmap(QtGui.QPixmap(self.path + "/images/logo.png"), QtGui.QIcon.Mode.Normal, QtGui.QIcon.State.Off)
         Dialog.setWindowIcon(icon)
         Dialog.setStyleSheet("QDialog {\n"
 "    background-color: #333333;\n"