+++ ScreenshotsParseTool/scripts/test.py
@@ -1,0 +1,4 @@
+import os
+
+
+print(os.path.join(os.path.expanduser('~'), 'Desktop'))